Did stress test, need anit depressants
So keeping a year worth of data sems not possible.
Maybe its to frequet data readings or the hardware is not strong enough.
The rPi literraly crashes when attempting to query data.
I use some of my gray matter to figure out a solution but i dont have high hopes.