Wrapped Sesnor Model Mock routine in try catch
In Order to recover wrap an error we would like to recover wrom and
error. This solution at the moment is flawed since we will ALWAYS
recover from the error. No the less, it is a simple fix. One should be
careful of adding other logic that might fail outside the try chatch
while loop.