Changed queries and JSON to be UTC
To avoide weird date time formats later in the consuming applications
and clients we made sure that all times are in UTC now. This is easier
since the rPi moves around and might change timezone without us knowing.
It is easier to let the client handle formatting of time.

Solves Item: 50